requirements.txt
PyJWT
coreapi
coreschema
djangorestframework
Django==2.2.6
Pillow==5.2.0
